second/requirements.txt
fastapi[all]
pydantic_settings
openai-agents[voice]